backend/requirements.txt
chromadb==0.6.3
fastapi==0.115.11
graphrag==2.1.0
langchain-core==0.3.45
langchain-text-splitters==0.3.7
openai==1.66.5
pydantic==2.10.6
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
tiktoken==0.8.0
uvicorn==0.34.0
celery==5.3.4
redis
minio==7.2.15
pdf2image==1.17.0
fitz
python-multipart
